+++ evalml/automl/automl_search.py
@@ -803,17 +803,18 @@ class AutoMLSearch:
         best = self.rankings.iloc[0]
         return self.get_pipeline(best["id"])
 
-    def save(self, file_path):
+    def save(self, file_path, pickle_protocol=cloudpickle.DEFAULT_PROTOCOL):
         """Saves AutoML object at file path
 
         Arguments:
             file_path (str): location to save file
+            pickle_protocol (int): the pickle data stream format.
 
         Returns:
             None
         """
         with open(file_path, 'wb') as f:
-            cloudpickle.dump(self, f)
+            cloudpickle.dump(self, f, protocol=pickle_protocol)
 
     @staticmethod
     def load(file_path):